+++ pkg/webui/account/containers/tokens-table/index.js
@@ -1,4 +1,4 @@
-// Copyright © 2020 The Things Network Foundation, The Things Industries B.V.
+// Copyright © 2022 The Things Network Foundation, The Things Industries B.V.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,23 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import status from '@ttn-lw/lib/store/logics/status'
+import TokensTable from './tokens-table'
+import connect from './connect'
 
-import init from './init'
-import user from './user'
-import identityServer from './identity-server'
-import sessions from './sessions'
-import clients from './clients'
-import collaborators from './collaborators'
-import authorizations from './authorizations'
+const ConnectedTokensTable = connect(TokensTable)
 
-export default [
-  ...status,
-  ...init,
-  ...user,
-  ...identityServer,
-  ...sessions,
-  ...clients,
-  ...collaborators,
-  ...authorizations,
-]
+export { ConnectedTokensTable as default, TokensTable }